+++ insights-ui/src/app/api/industry-tariff-reports/[industry]/generate-executive-summary/route.ts
@@ -1,4 +1,3 @@
-import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/00-industry-main-headings';
 import {
   getExecutiveSummaryAndSaveToFile,
   readExecutiveSummaryFromFile,
@@ -6,6 +5,7 @@ import {
 } from '@/scripts/industry-tariff-reports/01-executive-summary';
 import { readTariffUpdatesFromFile } from '@/scripts/industry-tariff-reports/03-industry-tariffs';
 import { getIndustryTariffReport, getSummariesOfEvaluatedAreas } from '@/scripts/industry-tariff-reports/industry-tariff-report-utils';
+import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
 import { IndustryTariffReport } from '@/scripts/industry-tariff-reports/tariff-types';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
