+++ insights-ui/src/app/api/industry-tariff-reports/[industry]/generate-headings/route.ts
@@ -1,12 +1,9 @@
 import { getIndustryTariffReport } from '@/scripts/industry-tariff-reports/industry-tariff-report-utils';
+import { readIndustryHeadingsFromFile, writeIndustryHeadingsToMarkdownFile } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
 import { IndustryTariffReport } from '@/scripts/industry-tariff-reports/tariff-types';
 import { NextRequest } from 'next/server';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
-import {
-  getAndWriteIndustryHeadings,
-  readIndustryHeadingsFromFile,
-  writeIndustryHeadingsToMarkdownFile,
-} from '@/scripts/industry-tariff-reports/00-industry-main-headings';
+import { getAndWriteIndustryHeadings } from '@/scripts/industry-tariff-reports/00-industry-main-headings';
 
 async function postHandler(req: NextRequest, { params }: { params: Promise<{ industry: string }> }): Promise<IndustryTariffReport> {
   const { industry } = await params;
