+++ insights-ui/src/app/api/industry-tariff-reports/[industry]/generate-industry-areas/route.ts
@@ -1,4 +1,5 @@
 import { getIndustryTariffReport } from '@/scripts/industry-tariff-reports/industry-tariff-report-utils';
+import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
 import { IndustryTariffReport } from '@/scripts/industry-tariff-reports/tariff-types';
 import { NextRequest } from 'next/server';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
@@ -7,7 +8,6 @@ import {
   readIndustryAreaSectionFromFile,
   writeIndustryAreaSectionToMarkdownFile,
 } from '@/scripts/industry-tariff-reports/05-industry-areas';
-import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/00-industry-main-headings';
 
 async function postHandler(req: NextRequest, { params }: { params: Promise<{ industry: string }> }): Promise<IndustryTariffReport> {
   const { industry } = await params;
