+++ insights-ui/src/app/api/industry-tariff-reports/[industry]/generate-tariff-updates/route.ts
@@ -1,4 +1,5 @@
 import { getIndustryTariffReport } from '@/scripts/industry-tariff-reports/industry-tariff-report-utils';
+import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
 import { IndustryTariffReport } from '@/scripts/industry-tariff-reports/tariff-types';
 import { NextRequest } from 'next/server';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
@@ -7,7 +8,6 @@ import {
   readTariffUpdatesFromFile,
   writeTariffUpdatesToMarkdownFile,
 } from '@/scripts/industry-tariff-reports/03-industry-tariffs';
-import { readIndustryHeadingsFromFile } from '@/scripts/industry-tariff-reports/00-industry-main-headings';
 
 interface GenerateTariffUpdatesRequest {
   date: string;
